+++ Libraries/Fields/Relationships/BelongsTo.php
@@ -28,7 +28,10 @@ class BelongsTo extends Relationship {
 	{
 		parent::__construct($field, $info, $config);
 
-		$relationship = $config->model->{$field}();
+		//set up the model depending on what's passed in
+		$model = is_a($config, 'Admin\\Libraries\\ModelConfig') ? $config->model : $config;
+
+		$relationship = $model->{$field}();
 		$otherModel = $relationship->model;
 
 		$this->table = $otherModel->table();