+++ Libraries/Fields/Relationships/HasManyAndBelongsTo.php
@@ -32,6 +32,7 @@ class HasManyAndBelongsTo extends Relationship {
 	{
 		parent::__construct($field, $info, $config);
 
+		//set up the model depending on what's passed in
 		$model = is_a($config, 'Admin\\Libraries\\ModelConfig') ? $config->model : $config;
 
 		$relationship = $model->{$field}();